Move old visualization scripts to v1 and other import assorting
