+++ micro_sam/util.py
@@ -1,32 +1,36 @@
-"""
-Helper functions for downloading Segment Anything models and predicting image embeddings.
+"""Helper functions for downloading Segment Anything models and predicting image embeddings.
 """
 
 import os
-import multiprocessing as mp
+import pooch
+import xxhash
 import pickle
 import hashlib
 import warnings
-from concurrent import futures
 from pathlib import Path
+import multiprocessing as mp
+from concurrent import futures
 from collections import OrderedDict
 from typing import Any, Dict, Iterable, List, Optional, Tuple, Union, Callable
 
-import elf.parallel as parallel_impl
-import imageio.v3 as imageio
 import numpy as np
-import pooch
+import imageio.v3 as imageio
+
 import segment_anything.utils.amg as amg_utils
-import torch
-import xxhash
+
 import zarr
 
 from elf.io import open_file
+import elf.parallel as parallel_impl
+
 from bioimage_cpp.utils import Blocking
 from bioimage_cpp.distance import distance_transform
 from bioimage_cpp.segmentation import relabel_sequential
+
 from skimage.measure import regionprops
 from skimage.segmentation import find_boundaries
+
+import torch
 from torchvision.ops.boxes import batched_nms
 
 from .__version__ import __version__
@@ -1782,8 +1786,9 @@ def mask_data_to_segmentation(
     """Convert the output of the automatic mask generation to an instance segmentation.
 
     Args:
-        masks: The outputs generated by `AutomaticMaskGenerator`, other classes from `micro_sam.v1.instance_segmentation`,
-            or from `micro_sam.v1.inference` functions. Only supported for output_mode=binary_mask.
+        masks: The outputs generated by `AutomaticMaskGenerator`, other classes from
+            `micro_sam.v1.instance_segmentation`, or from `micro_sam.v1.inference` functions. Only
+            supported for output_mode=binary_mask.
         shape: The shape of the output segmentation. If None, it will be derived from the mask input.
             If the mask where predicted with tiling then the shape must be given.
         min_object_size: The minimal size of an object in pixels. By default, set to '0'.
